# api/swagger/swagger.json
{
  "info": {
    "title": "Security Service",
    "description": "This service is responsible for security",
    "version": "1.0.0"
  },
  "swagger": "2.0",
  "paths": {
    "/health": {
      "get": {
        "description": "Get Health of Service",
        "tags": [
          "health"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Health",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Health"
            }
          }
        }
      }
    },
    "/security/oauth": {
      "get": {
        "description": "Request Authorization",
        "operationId": "authorise",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-client-id",
            "description": "Client Id",
            "in": "header",
            "required": true
          },
          {
            "name": "x-client-secret",
            "description": "Client Secret",
            "in": "header",
            "required": true
          },
          {
            "name": "grant_type",
            "description": "Grant Type (i.e. token)",
            "in": "query",
            "required": true
          },
          {
            "name": "scope",
            "description": "Scope",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "AuthorizationResponse",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AuthorizationResponse"
            }
          },
          "401": {
            "description": "Error",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/security/oauth/token": {
      "get": {
        "description": "Request Token",
        "operationId": "token",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "tokenHashAccessCode",
            "description": "Token Hash Access Code",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "TokenResponse",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/TokenResponse"
            }
          },
          "404": {
            "description": "Error",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/security/oauth/token/_check": {
      "get": {
        "description": "Check Token Validity",
        "operationId": "check",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "access-token",
            "description": "accessToken",
            "in": "header",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "TokenValidity",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/AccessTokenValidity"
            }
          },
          "404": {
            "description": "Error",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Error": {
      "type": "object",
      "required": [
        "errorMessage"
      ],
      "properties": {
        "errorMessage": {
          "type": "string"
        }
      }
    },
    "Health": {
      "type": "object",
      "required": [
        "cpuPercentUsage",
        "totalMemPercentageUsage",
        "loadAvg"
      ],
      "properties": {
        "cpuPercentUsage": {
          "type": "number"
        },
        "totalMemPercentageUsage": {
          "type": "number"
        },
        "loadAvg": {
          "type": "number"
        }
      }
    },
    "AccessTokenValidity": {
      "type": "object",
      "required": [
        "valid"
      ],
      "properties": {
        "valid": {
          "type": "boolean"
        }
      }
    },
    "AuthorizationResponse": {
      "type": "object",
      "required": [
        "redirect_uri"
      ],
      "properties": {
        "redirect_uri": {
          "type": "string"
        }
      }
    },
    "TokenResponse": {
      "type": "object",
      "required": [
        "access_token"
      ],
      "properties": {
        "access_token": {
          "type": "string"
        }
      }
    }
  },
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": [
    {
      "name": "error",
      "description": "Everything you need to know about Error"
    },
    {
      "name": "health",
      "description": "Everything you need to know about Health"
    },
    {
      "name": "authorizationResponses",
      "description": "Everything you need to know about AuthorizationResponses"
    },
    {
      "name": "tokenResponses",
      "description": "Everything you need to know about TokenResponses"
    }
  ]
}